feat: 增加pdf图片和文字解析以及代码整理

+++ code/main.py
@@ -18,13 +18,10 @@ usage:
     cd code
     python main.py # 无参数默认扫描"../data"文件夹
 """
-# 计时
-T1 = time.perf_counter()
 
 # 添加日志模块
 logger = LoggerSingleton().get_logger()
 TAG = "main.py: "
-logger.info(TAG + "************************ start *****************************")
 
 # 添加结果输出模块
 res_out = ResOut()
@@ -49,6 +46,11 @@ argparse.add_argument("-t", "--time", default="time_info.txt",
 # 是否处理非图片文件内部中的图片, 默认为true
 argparse.add_argument("-p", "--picture", default="true",
                       help="process picture in non image files")
+
+# 程序运行结束后是否清空workspace缓存目录
+argparse.add_argument("-c", "--clean", default="true",
+                      help="clean the workspace dir in the end")
+
 
 # 输入扫描的路径
 args = argparse.parse_args()
@@ -91,6 +93,7 @@ direct_controller = fileUtil.DirectController()
 process_manager = processUtil.ProcessManager()
 
 
+################################## main() #####################################
 def main():
     # 逐个根目录执行
     while not globalVar.root_folder_list.empty():
@@ -119,6 +122,7 @@ def main():
             process_manager.release_process_pool()
     # Your code here
 
+
 # #性能分析工具下的执行
 # profiler = cProfile.Profile()
 # profiler.run('main()')
@@ -126,30 +130,41 @@ def main():
 # profiler.dump_stats('./log/profile_results.prof')
 
 
-# 不使用性能分析工具直接执行
+logger.info(TAG + "************************ start *****************************")
+T1 = time.perf_counter()  # 计时
+
 main()
 
+T2 = time.perf_counter()  # 计时结束
+logger.info(TAG + "************************* end ******************************")
+
+
 # 清空工作区
-command = "rm -rf ../workspace/*"
-result = subprocess.run(
-    command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
-if result.returncode == 0:
-    logger.info('工作区已清空')
+if args.clean == "true":
+    logger.info(TAG+"==>准备清空工作区")
+    command = "rm -rf ../workspace/*"
+    result = subprocess.run(
+        command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
+    if result.returncode == 0:
+        logger.info('工作区已清空')
+    else:
+        logger.info('工作区清空失败')
+        globalVar.flag_list.append(True)
 else:
-    logger.info('工作区清空失败')
+    logger.info(TAG+"==>不清空工作区!!")
 
 
 # 将结果写入文件
 output_tile_path = "output/"+datetime.now().strftime("%Y%m%d%H%M%S%f") + \
     "_output.json"
 res_out.save_to_file(output_tile_path)
-logger.info(TAG+"************************* end ******************************")
 logger.info(TAG+"result is saved to: "+output_tile_path+" , total is " +
-            str(len(res_out.res_json)-2) + " term")
+            str(len(res_out.res_json)) + " term")
 
-# 计时结束
-T2 = time.perf_counter()
+
+# 打印程序耗时
 logger.info(TAG+'程序运行时间:%s毫秒' % ((T2 - T1)*1000))
+
 
 # 记录程序运行时间
 time_info = TAG+'程序运行时间:%s毫秒' % ((T2 - T1)*1000)
